backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
huggingface_hub
